+++ webpack.config.js
@@ -1,10 +1,12 @@
 var webpack = require('webpack');
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var version = require('./package.json').version;
+var packageJson = require('./package.json');
+var version = packageJson.version;
 
 module.exports = {
     entry: {
-        'angular-popups': './src/index.js'
+        'angular-popups': './src/popups.js',
+        'angular-popups-nocss': './src/popups-nocss.js'
     },
     output: {
         path: 'dist',
@@ -13,41 +15,30 @@ module.exports = {
     devtool: 'source-map',
     plugins: [
         new ExtractTextPlugin('[name].css'),
-        new webpack.BannerPlugin('angular-popups@' + version + ' | https://github.com/aui/angular-popups')
+        new webpack.BannerPlugin(packageJson.name + '@' + packageJson.version + ' | ' + packageJson.homepage)
     ],
     externals: {
         angular: 'angular'
     },
     module: {
-        preLoaders: [
-            {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                loader: 'jshint-loader'
-            }
-        ],
+        preLoaders: [{
+            test: /\.js$/,
+            exclude: /node_modules/,
+            loader: 'jshint-loader'
+        }],
         loaders: [
             //{test: /\.css$/, loader: ExtractTextPlugin.extract('style-loader', 'css-loader')}, // 把 CSS 抽离成独立的文件
-            {test: /\.css$/, loader: 'style!css?minimize'}
+            {
+                test: /\.css$/,
+                loader: 'style!css?sourceMap&minimize'
+            }
         ]
     },
 
-    // more options in the optional jshint object
     jshint: {
-        // any jshint option http://www.jshint.com/docs/options/
-        // i. e.
         camelcase: true,
-
-        // jshint errors are displayed by default as warnings
-        // set emitErrors to true to display them as errors
         emitErrors: false,
-
-        // jshint to not interrupt the compilation
-        // if you want any file with jshint errors to fail
-        // set failOnHint to true
         failOnHint: false,
-
-        // custom reporter function
-        reporter: function(errors) { }
+        reporter: function(errors) {}
     }
 };